Added Map Component
